+++ openmetadata-ui/src/main/resources/ui/src/components/Users/Users.util.tsx
@@ -1,6 +1,18 @@
+import { Popover, Space, Tag } from 'antd';
+import { ColumnsType } from 'antd/lib/table';
+import { t } from 'i18next';
+import { isEmpty, isUndefined, uniqueId } from 'lodash';
 import React from 'react';
+import { Link } from 'react-router-dom';
+import { getUserPath } from '../../constants/constants';
 import { User } from '../../generated/entity/teams/user';
 import { EntityReference } from '../../generated/type/entityReference';
+import { getEntityName } from '../../utils/CommonUtils';
+import { LIST_CAP } from '../../utils/PermissionsUtils';
+import {
+  getRoleWithFqnPath,
+  getTeamsWithFqnPath,
+} from '../../utils/RouterUtils';
 import SVGIcons, { Icons } from '../../utils/SvgUtils';
 
 export const userPageFilterList = [
@@ -67,3 +79,115 @@ export const getUserFromEntityReference = (entity: EntityReference): User => {
     email: '',
   };
 };
+
+export const commonUserDetailColumns = (): ColumnsType<User> => [
+  {
+    title: t('label.username'),
+    dataIndex: 'username',
+    key: 'username',
+    render: (_, record) => (
+      <Link
+        className="hover:tw-underline tw-cursor-pointer"
+        data-testid={record.name}
+        to={getUserPath(record.fullyQualifiedName || record.name)}>
+        {getEntityName(record)}
+      </Link>
+    ),
+  },
+  {
+    title: t('label.teams'),
+    dataIndex: 'teams',
+    key: 'teams',
+    render: (_, record) => {
+      const listLength = record.teams?.length ?? 0;
+      const hasMore = listLength > LIST_CAP;
+
+      if (isUndefined(record.teams) || isEmpty(record.teams)) {
+        return <>No Team</>;
+      } else {
+        return (
+          <Space wrap data-testid="policy-link" size={4}>
+            {record.teams.slice(0, LIST_CAP).map((team) => (
+              <Link
+                className="hover:tw-underline tw-cursor-pointer"
+                key={uniqueId()}
+                to={getTeamsWithFqnPath(team.fullyQualifiedName ?? '')}>
+                {getEntityName(team)}
+              </Link>
+            ))}
+            {hasMore && (
+              <Popover
+                className="tw-cursor-pointer"
+                content={
+                  <Space wrap size={4}>
+                    {record.teams.slice(LIST_CAP).map((team) => (
+                      <Link
+                        className="hover:tw-underline tw-cursor-pointer"
+                        key={uniqueId()}
+                        to={getTeamsWithFqnPath(team.fullyQualifiedName ?? '')}>
+                        {getEntityName(team)}
+                      </Link>
+                    ))}
+                  </Space>
+                }
+                overlayClassName="tw-w-40 tw-text-center"
+                trigger="click">
+                <Tag className="tw-ml-1" data-testid="plus-more-count">{`+${
+                  listLength - LIST_CAP
+                } more`}</Tag>
+              </Popover>
+            )}
+          </Space>
+        );
+      }
+    },
+  },
+  {
+    title: t('label.roles'),
+    dataIndex: 'roles',
+    key: 'roles',
+    render: (_, record) => {
+      const listLength = record.roles?.length ?? 0;
+      const hasMore = listLength > LIST_CAP;
+
+      if (isUndefined(record.roles) || isEmpty(record.roles)) {
+        return <>No Role</>;
+      } else {
+        return (
+          <Space wrap data-testid="policy-link" size={4}>
+            {record.roles.slice(0, LIST_CAP).map((role) => (
+              <Link
+                className="hover:tw-underline tw-cursor-pointer"
+                key={uniqueId()}
+                to={getRoleWithFqnPath(role.fullyQualifiedName ?? '')}>
+                {getEntityName(role)}
+              </Link>
+            ))}
+            {hasMore && (
+              <Popover
+                className="tw-cursor-pointer"
+                content={
+                  <Space wrap size={4}>
+                    {record.roles.slice(LIST_CAP).map((role) => (
+                      <Link
+                        className="hover:tw-underline tw-cursor-pointer"
+                        key={uniqueId()}
+                        to={getRoleWithFqnPath(role.fullyQualifiedName ?? '')}>
+                        {getEntityName(role)}
+                      </Link>
+                    ))}
+                  </Space>
+                }
+                overlayClassName="tw-w-40 tw-text-center"
+                trigger="click">
+                <Tag className="tw-ml-1" data-testid="plus-more-count">{`+${
+                  listLength - LIST_CAP
+                } more`}</Tag>
+              </Popover>
+            )}
+          </Space>
+        );
+      }
+    },
+  },
+];
